requirements.txt
openai-whisper
ollama
python-dotenv
google-genai
torch
numpy
psycopg2-binary
